chore: remove legacy configuration files and update project structure for Chrome extension
