+++ packages/chrome/.env.example
@@ -1,0 +1,3 @@
+# ElevenLabs API Configuration
+# Get your API key from: https://elevenlabs.io/app/settings/api-keys
+VITE_ELEVENLABS_API_KEY=your_elevenlabs_api_key_here
