+++ packages/chrome/README.md
