+++ packages/chrome/privacy-policy.md
