+++ packages/chrome/public/manifest.json
@@ -1,0 +1,24 @@
+{
+	"manifest_version": 3,
+	"name": "Nihongo Speech",
+	"version": "1.0.0",
+	"description": "Convert Japanese sentences to realistic male and female speech using advanced TTS APIs.",
+	"icons": {
+		"16": "icon-16.png",
+		"48": "icon-48.png",
+		"128": "icon-128.png"
+	},
+	"action": {
+		"default_popup": "popup.html",
+		"default_icon": {
+			"16": "icon-16.png",
+			"48": "icon-48.png",
+			"128": "icon-128.png"
+		}
+	},
+	"background": {
+		"service_worker": "background.js"
+	},
+	"permissions": ["storage"],
+	"host_permissions": ["https://api.elevenlabs.io/*"]
+}
