+++ packages/chrome/src/vite-env.d.ts
@@ -1,9 +1,9 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
-  readonly VITE_ELEVENLABS_API_KEY: string
+	readonly VITE_ELEVENLABS_API_KEY: string;
 }
 
 interface ImportMeta {
-  readonly env: ImportMetaEnv
-} 
+	readonly env: ImportMetaEnv;
+}
