--- public/icon.png
@@ -1,1 +1,0 @@
-// Placeholder: Add your 128x128 icon.png file here for the Chrome extension icon.
