--- src/__tests__/setupTests.ts
@@ -1,1 +1,0 @@
-import '@testing-library/jest-dom';
