--- src/popup/index.tsx
@@ -1,5 +1,0 @@
-import { createRoot } from 'react-dom/client';
-import Popup from './Popup';
-
-const root = createRoot(document.getElementById('root')!);
-root.render(<Popup />);
