--- src/popup/popup.css
@@ -1,200 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@font-face {
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url('/fonts/inter-variable.ttf') format('truetype');
-}
-
-:root {
-    --primary: rgba(253, 255, 121, 1);
-    --secondary: rgba(20, 22, 18, 0.5);
-}
-
-html {
-  background-color: var(--secondary);
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  border-radius: 23px;
-}
-
-body {
-  font-family: 'Inter', sans-serif;
-  background-color: var(--secondary);
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  border-radius: 23px;
-}
-
-#root {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-*, *::before, *::after {
-  box-sizing: border-box;
-}
-
-input, button {
-    font-family: 'Inter', sans-serif;
-}
-
-.custom-radio {
-  display: inline-flex;
-  align-items: center;
-  cursor: pointer;
-  margin-right: 6px;
-  font-size: 16px;
-  user-select: none;
-}
-
-.custom-radio input[type="radio"] {
-  opacity: 0;
-  position: absolute;
-  width: 0;
-  height: 0;
-}
-
-.custom-radio .radio-checkmark {
-  position: relative;
-  height: 20px;
-  width: 20px;
-  background-color: var(--secondary);
-  border: 2px solid #ccc;
-  border-radius: 50%;
-  margin-right: 4px;
-  transition: all 0.2s ease-out;
-}
-
-.custom-radio input[type="radio"]:focus + .radio-checkmark {
-  box-shadow: 0 0 0 1px var(--primary);
-}
-
-.custom-radio:hover .radio-checkmark {
-  border-color: #aaa;
-}
-
-.custom-radio .radio-checkmark::after {
-  content: "";
-  position: absolute;
-  display: none;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background-color: var(--primary);
-}
-
-.custom-radio input[type="radio"]:checked + .radio-checkmark::after {
-  display: block;
-}
-
-.custom-radio input[type="radio"]:checked + .radio-checkmark {
-  border-color: var(--primary);
-}
-
-audio {
-  width: 100%;
-  height: 40px;
-  border-radius: 12px;
-  background-color: var(--secondary);
-  outline: none;
-}
-
-audio::-webkit-media-controls-panel {
-  background-color: var(--secondary);
-  border-radius: 12px;
-}
-
-audio::-webkit-media-controls-play-button {
-  background-color: var(--primary);
-  border-radius: 50%;
-}
-
-audio::-webkit-media-controls-current-time-display,
-audio::-webkit-media-controls-time-remaining-display {
-  color: white;
-}
-
-audio::-webkit-media-controls-timeline {
-  background-color: #333;
-  border-radius: 4px;
-}
-
-audio::-webkit-media-controls-volume-slider {
-  background-color: var(--primary);
-}
-
-.audio-progress,
-.audio-volume {
-  -webkit-appearance: none;
-  appearance: none;
-  height: 8px;
-  border-radius: 4px;
-  cursor: pointer;
-  outline: none;
-}
-
-.audio-progress::-webkit-slider-track,
-.audio-volume::-webkit-slider-track {
-  height: 8px;
-  border-radius: 4px;
-  border: none;
-  background: transparent;
-}
-
-.audio-progress::-webkit-slider-thumb,
-.audio-volume::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  height: 16px;
-  width: 12px;
-  border-radius: 0;
-  background: var(--primary);
-  cursor: pointer;
-  border: none;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-  position: relative;
-}
-
-.audio-progress::-webkit-slider-thumb:hover,
-.audio-volume::-webkit-slider-thumb:hover {
-  background: var(--primary);
-  transform: scale(1.1);
-}
-
-.audio-progress::-moz-range-track,
-.audio-volume::-moz-range-track {
-  background: rgba(253, 255, 121, 0.1);
-  height: 8px;
-  border-radius: 4px;
-  border: none;
-}
-
-.audio-progress::-moz-range-thumb,
-.audio-volume::-moz-range-thumb {
-  height: 16px;
-  width: 16px;
-  border-radius: 50%;
-  background: var(--primary);
-  cursor: pointer;
-  border: 2px solid var(--secondary);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-  -moz-appearance: none;
-}
-
-.audio-progress::-moz-range-progress,
-.audio-volume::-moz-range-progress {
-  background: var(--primary);
-  height: 8px;
-  border-radius: 4px;
-}
