--- tailwind.config.js
@@ -1,18 +1,0 @@
-module.exports = {
-  content: [
-    './src/**/*.{js,ts,jsx,tsx,html}',
-    './public/*.html'
-  ],
-  theme: {
-    extend: {
-    colors: {
-      'primary': 'rgba(253, 255, 121, 1)',
-      'secondary': 'rgba(20, 22, 18, 1)',
-    },
-    fontFamily: {
-      'mono': ['Berkeley Mono','Fira Code','ui-monospace', 'monospace'],
-    },
-   },
-  },
-  plugins: [],
-};
